telegram-bots/requirements.txt
# PicoClaw Bot Handlers — Python dependencies
# Install: pip install -r requirements.txt

python-telegram-bot==20.*
pyyaml>=6.0
inotify-simple>=1.3
flask>=3.0

# Testing
pytest>=7.0
pytest-asyncio>=0.21
